return map idea test

+++ src/main/java/com/mycollectreport/service/ElasticsearchService.java
@@ -4,5 +4,5 @@ import java.io.IOException;
 import java.util.Map;
 
 public interface ElasticsearchService {
-    Map<String, Integer> getCounts() throws IOException;
+    Map<String, Long> getCounts() throws IOException;
 }
